#60feoAnA: Moved .editorconfig from config/ to root directory to make it work. Executed ESLint fix & prettier format. Refactored PawsTitle (now PawsHeading). Added styling for storybook only to meet Frontend application design.
