+++ env.d.ts
@@ -1,6 +1,6 @@
 /// <reference types="vite/client" />
 
 declare module "*.module.css" {
-  const classes: { [key: string]: string };
-  export default classes;
+	const classes: { [key: string]: string };
+	export default classes;
 }
