--- src/assets/fonts.css
@@ -1,7 +1,0 @@
-@font-face {
-  font-family: "Wix Madefor Display";
-  src: url("./fonts/WixMadeforDisplay-Variable.ttf")
-    format("truetype-variations");
-  font-weight: 400 800;
-  font-style: normal;
-}
