+++ src/components/Icon/icon.ts
@@ -1,0 +1,5 @@
+export interface BaseIcon {
+	color?: string;
+	width?: string;
+	height?: string;
+}
