+++ src/components/PawsButton/PawsButton.module.css
@@ -1,27 +1,27 @@
 .pawsButton {
-  font-family: var(--paws-font-family);
-  border: none;
-  color: var(--paws-color-text-on-accent);
-  padding: 10px 18px;
-  border-radius: 8px;
-  font-weight: 500;
-  cursor: pointer;
-  transition: filter 0.2s;
-  &:hover {
-    filter: brightness(1.15);
-  }
-  &.primary {
-    background-color: var(--paws-color-accent-primary);
-    font-weight: bold;
-  }
-  &.secondary {
-    background-color: var(--paws-color-accent-secondary);
-  }
-  &.small {
-    height: 20px;
-    font-size: 12px;
-    font-weight: 600;
-    padding: 0 10px;
-    border-radius: 4px;
-  }
+	font-family: var(--paws-font-family);
+	border: none;
+	color: var(--paws-color-text-on-accent);
+	padding: 10px 18px;
+	border-radius: 8px;
+	font-weight: 500;
+	cursor: pointer;
+	transition: filter 0.2s;
+	&:hover {
+		filter: brightness(1.15);
+	}
+	&.primary {
+		background-color: var(--paws-color-accent-primary);
+		font-weight: bold;
+	}
+	&.secondary {
+		background-color: var(--paws-color- accent-secondary);
+	}
+	&.small {
+		height: 20px;
+		font-size: 12px;
+		font-weight: 600;
+		padding: 0 10px;
+		border-radius: 4px;
+	}
 }
