+++ src/components/PawsButton/PawsButton.vue
@@ -1,7 +1,7 @@
 <template>
-  <button :class="[styles.pawsButton, buttonClass]">
-    {{ label }}
-  </button>
+	<button :class="[styles.pawsButton, buttonClass]">
+		{{ label }}
+	</button>
 </template>
 
 <script setup lang="ts">
@@ -9,25 +9,25 @@ import { computed } from "vue";
 
 import styles from "./PawsButton.module.css";
 const props = defineProps({
-  label: {
-    type: String,
-    required: true,
-  },
-  variant: {
-    type: String,
-    default: "primary",
-  },
-  size: {
-    type: String,
-    default: "normal",
-  },
+	label: {
+		type: String,
+		required: true,
+	},
+	variant: {
+		type: String,
+		default: "primary",
+	},
+	size: {
+		type: String,
+		default: "normal",
+	},
 });
 const buttonClass = computed(() => {
-  return {
-    [styles.primary]: props.variant === "primary",
-    [styles.secondary]: props.variant === "secondary",
-    [styles.small]: props.size === "small",
-  };
+	return {
+		[styles.primary]: props.variant === "primary",
+		[styles.secondary]: props.variant === "secondary",
+		[styles.small]: props.size === "small",
+	};
 });
 </script>
