--- src/components/PawsTitle/PawsTitle.module.css
@@ -1,11 +1,0 @@
-.pawsTitle {
-  font-family: var(--paws-font-family);
-  color: var(--paws-color-text-primary);
-  font-size: 28px;
-  font-weight: 600;
-}
-
-.subtitle {
-  font-size: 16px;
-  font-weight: 500;
-}
