--- src/themes/paws-dark.css
@@ -1,17 +1,0 @@
-.dark-theme {
-  --paws-color-accent-primary: #b4355d;
-  --paws-color-accent-secondary: #62223e;
-  --paws-color-text-primary: #ffffff;
-  --paws-color-text-secondary: #d8d8d8;
-  --paws-color-text-on-accent: #ffffff;
-  --paws-color-success: #4caf50;
-  --paws-color-warning: #ffc107;
-  --paws-color-error: #f44336;
-  --paws-color-bg-darkest: #222222;
-  --paws-color-bg-dark: #262626;
-  --paws-color-bg-lighter: #383838;
-  --paws-color-bg-disabled: #4a4a4a;
-  --paws-color-text-disabled: #8a8a8a;
-  --paws-color-button-disabled: #3e3e3e;
-  --paws-font-family: "Wix Madefor Display", sans-serif;
-}
